% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_curve.R
\name{accum_curve}
\alias{accum_curve}
\title{Accumulation curve of network dimensions}
\usage{
accum_curve(
  int_data,
  property = c("vcount", "ecount", "edge_density"),
  k = 100
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{property}{\strong{property}: indicates the network property, obtained from igraph
functions, which accuracy is being evaluated. Only three options are
currently available:
\itemize{
\item \emph{vcount}: accuracy of the estimated number of nodes (using igraph function
"vcount").
\item \emph{ecount}: accuracy of the estimated number of canopy-recruit interactions
(using igraph function "ecount").
\item \emph{edge_density}: accuracy of the estimated network connectance (using
igraph function "edge_density).
}}

\item{k}{An integer number specifying the number of random repetitions of subsets of
\emph{n} plots. In each of the \emph{k} repetitions, a subset of \emph{n} randomly chosen
plots is combined to build a partial network for which the indicated
property is estimated. High values provide more confident estimates of the
accuracy, but the function may take long time if \emph{k} >> 100.}
}
\value{
The function returns a list of two objects:
\itemize{
\item A plot representing the mean and 95\% Confidence interval (i.e. 1.96 times
the standard error) of the estimate of the property selected when an
increasing number of randomly selected plots are considered.
\item A data frame with the cumulative values of the property for each
repetition of \emph{k} plots. Provided so you can prepare your own customized
accumulation plot.
}
}
\description{
Uses plot accumulation curves to assess whether the estimates of the number
of nodes, links, and link density (a.k.a. connectance) are stable (see Pulgar
et al. 2017 for other descriptors). The function accum_curve plots
accumulation curves for parameters that can be obtained with package
\strong{\code{igraph}} (Csardi & Nepusz, 2006).

\strong{NOTE}: This function is intended for data sets organised in multiple
plots of the same community or locality.
}
\examples{
accum_links <- accum_curve(Amoladeras_int, property="ecount", k=10)
head(accum_links$Data)
accum_links$Plot
}
