% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_cover.R
\name{aggr_cover}
\alias{aggr_cover}
\title{Calculate aggregated species cover}
\usage{
aggr_cover(dataset, site)
}
\arguments{
\item{dataset}{character. Name of the data set where cover data are stored.
Usually, it will be "CanopyCover" (or the name assigned to the
dataset "CanopyCover.csv" when it was imported).}

\item{site}{character. Name of a study site.}
}
\value{
A data frame containing the total cover of each species in a local community,
aggregated across plots. It has two columns: "Canopy", with the name of each species (irrespective
of being acting as canopies or recruits); and "abundance".
}
\description{
Calculate the aggregated (i.e. summed) cover of each species across
all the plots from a local community.
}
\examples{
#Ventisquero_cover <- aggr_cover(CanopyCover, site = "Ventisquero")
}
