% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr_RN.R
\name{aggr_RN}
\alias{aggr_RN}
\title{Prepare recruitment data}
\usage{
aggr_RN(raw_RN)
}
\arguments{
\item{raw_RN}{A data frame containing, at least, the columns:
"Canopy", "Recruit" and "Frequency".}
}
\value{
A data frame containing the data to build a recruitment network
of a local community: "canopy", "recruit", "fij", "Tij" and "Pij".
}
\description{
This function makes a data frame with the basic data needed to build the
recruitment network: the identity of canopy and recruit species, and several
possible weighting variables for the interactions: fij is the number of
recruits of species i interacting with canopy species j summed across plots,
Tij is the number of plots (i.e. incidence) where the interaction has been found,
and Pij is the presence/absence of the interaction in the whole study site.

The output contains all possible pairs of interacting species,
including those that were not observed, which are assigned weight = 0.
}
\examples{
#Ventisquero_RN <- aggr_RN(Ventisquero_raw)
}
