% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Amoladeras_cover}
\alias{Amoladeras_cover}
\title{Amoladeras cover data}
\format{
An object of class \code{data.frame} with 282 rows and 4 columns.
}
\usage{
Amoladeras_cover
}
\description{
Amoladeras cover data
}
\keyword{datasets}
