% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Amoladeras_int}
\alias{Amoladeras_int}
\title{Amoladeras interaction data}
\format{
An object of class \code{data.frame} with 824 rows and 4 columns.
}
\usage{
Amoladeras_int
}
\description{
Amoladeras interaction data
}
\keyword{datasets}
