% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asocindex.R
\name{asocindex}
\alias{asocindex}
\title{TODO: Title}
\usage{
asocindex(inter = com[[1]], Canopy_all = com[[2]], adjlist = com[[3]])
}
\arguments{
\item{inter}{a data frame which must include at least four columns named:
"Study_site", "Recruit", "Canopy", "Freq", with Freq being the number of
recruits of a given species under each canopy species (including "Open"
as a canopy species if the recruit is on the open (i.e. no canopy)),
in each study site. It can be obtained extracting the first element of the
list generated by the function plot_to_com.}

\item{Canopy_all}{a data frame which must include at least three columns named:
"Study_site", "Canopy", "Canopy_cover", with the last being the \% of cover
of each canopy species (or Open) in each study site. It can be be obtained
extracting the second element of the list generated by the function plot_to_com.}

\item{adjlist}{a list with the community matrix with recruit and canopy species
in rows and columns respectively in each study site. It can be be obtained
extracting the third element of the list generated by the function plot_to_com.}
}
\value{
a data frame with the number of recruits observed under each canopy
species and in the open, with the relative cover of each canopy species and
Open in each study site.
}
\description{
Generate a data frame with the information required to calculate
association indices, with the number of recruits observed under each canopy
species and in the open for multiple study site.
The function removes recruit species with a single individual.
}
\examples{
#com <- plot_to_com(data, dbcover)
#db_asoc <- asocindex(com[[1]], com[[2]], com[[3]])
}
