% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_service_test.R
\name{canopy_service_test}
\alias{canopy_service_test}
\title{Identifies whether a species acts as depressing, enhancing or neutral}
\usage{
canopy_service_test(int_data, cover_data)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}
}
\value{
Data frame with the same variables provided by the function
\code{\link[=int_significance]{int_significance()}} but without distinguishing between recruit species.
}
\description{
Tests whether a canopy species has a depressing, enhancing or neutral
effect on recruitment in general (i.e., at the community level) compared
to the "Open", independently considering all the recruit species together.
}
\examples{
canopy_service_test(Amoladeras_int, Amoladeras_cover)

}
