% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CanopyCover}
\alias{CanopyCover}
\title{RecruitNet-Cover}
\format{
An object of class \code{data.frame} with 26812 rows and 7 columns.
}
\source{
https://doi.org/10.1002/ecy.3923
}
\usage{
CanopyCover
}
\description{
Canopy cover of plant-plant interactions global database
}
\keyword{datasets}
