% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interactions.R
\name{check_interactions}
\alias{check_interactions}
\title{Check interaction data format}
\usage{
check_interactions(int_data = NULL)
}
\arguments{
\item{int_data}{data frame containing interaction data.}
}
\value{
The function will return error(s) if problems are detected. Otherwise
an OK message.
}
\description{
A canopy-recruit interactions data set must contain, at least, the raw data of the observed frequency of each canopy-recruit pair in each sampling unit (e.g. plot, quadrat, transect). For standardization, the columns in this data set must be named:
\itemize{
\item \strong{Plot}. Plot ID. Uniquely identifies each of the sampling units over which the interactions have been surveyed. Include this column even if you only surveyed one plot.
\item \strong{Canopy}. Name of the canopy species. If your survey included any type of open interspaces (e.g., canopy gaps, open ground, spaces away from a canopy plant), include these as a single node named "Open" (note the capital "O"). If you use full scientific names (e.g. latin binomials), concatenate the epithets with a lower dash (e.g. Olea_europaea).
\item \strong{Recruit}. Name of the recruit species. If you use full scientific names, concatenate the epithets with a lower dash (e.g. Olea_europaea).
\item \strong{Frequency}. Frequency of the canopy-recruit interaction in the study plot, indicated as number of individuals of the recruit species found under individuals of the canopy species.
}
}
\examples{
int_data <- load_interactions()
check_interactions(int_data)
}
