% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_subset.R
\name{comm_subset}
\alias{comm_subset}
\title{Subset sites}
\usage{
comm_subset(dataset = NULL, site = NULL)
}
\arguments{
\item{dataset}{A data frame containing recruitment data.}

\item{site}{character. Name of a study site within the dataset or
a character vector with the names of several study sites.}
}
\value{
A data frame containing all the data from the desired community/ies
}
\description{
This function extracts the information of a given local community and
stores it in a separate data frame. It can also extract more than one site.
}
\examples{
data(RecruitNet)

Ventisquero <- comm_subset(RecruitNet, site = "Ventisquero")

TwoSites <- comm_subset(RecruitNet, c("Ventisquero","Agadir"))
}
