% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_to_RN.R
\name{comm_to_RN}
\alias{comm_to_RN}
\title{Combine cover data with interaction data.}
\usage{
comm_to_RN(
  int_data,
  cover_data,
  expand = c("yes", "no"),
  rm_sp_no_cover = c("allsp", "onlycanopy")
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{expand}{Indicates whether to expand the dataframe to include the non-detected
interactions between every pair of species that occurs in the study site,
assigning values of 0 to their frequency. It can take two possible values:
\itemize{
\item \emph{yes}: Generates all possible interactions between canopy and recruit
species in the study site (i.e. detected and non-detected), adding values
of interaction strength of 0 to the non-detected interactions.
\item \emph{no}: Considers only the pairwise interactions that were actually
detected in the study site.
}}

\item{rm_sp_no_cover}{Defines the filtering behavior, specifying which species should be removed
based on the availability of cover data. It can take two options:
\itemize{
\item \emph{allsp}: removes all species lacking cover data (both canopy and recruit
species).
\item \emph{onlycanopy}: removes only canopy species lacking cover data, retaining
recruit species even if they lack cover data.
}}
}
\value{
Data frame with the following information for each pair of species:
\itemize{
\item \emph{Canopy}. Name of the canopy species, including "Open" interspaces. If you
use full scientific names, concatenate the epithets with a lower dash (e.g.
Olea_europaea).
\item \emph{Recruit}. Name of the recruit species. If you use full scientific names,
concatenate the epithets with a lower dash (e.g. Olea_europaea).
\item \emph{Ac}: Area (in m^2^), (or m when relative cover is measured with transects)
occupied by the canopy species (or "Open" interspaces) in the total area
sampled in the study site.
\item \emph{Ar}. Area (in m^2^), (or m when relative cover is measured with transects)
occupied by the recruit species (observed as adults), in the total area
sampled in the study site.
\item \emph{Fcr}: frequency of recruitment, in number of recruits by canopy-recruit
pair.
\item \emph{Icr}: incidence of the interaction, as number of plots where it was
observed.
\item \emph{Pcr}: presence/absence (1/0) of the interaction in the study site. This
provides the "unweighted" version of the adjacency matrix.
}
}
\description{
Combine interactions and cover data sets from a study site into a single data
frame. For study sites where data has been collected in several plots, this
function collapses the information into a single value per pairwise interaction.
You can choose whether or not the output contains all possible pairs of
interacting species (including those that were not observed, which are
assigned weight = 0) and whether or not to remove species lacking information
on canopy cover.
}
\examples{
df <- comm_to_RN(Amoladeras_int, Amoladeras_cover, expand="yes",
rm_sp_no_cover="allsp")
head(df)

}
