% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_RN.R
\name{download_RN}
\alias{download_RN}
\title{Download RecruitNet database}
\usage{
download_RN(path = getwd(), destfile = "RN.zip", unzip = TRUE)
}
\arguments{
\item{path}{character. Path to folder where the RecruitNet database should be saved.}

\item{destfile}{character. Name of the zip file to be saved.}

\item{unzip}{Logical. Uncompress the zip file? Default is TRUE.}
}
\value{
A zip file or two CSV files, depending if unzip is TRUE.
}
\description{
Download latest version of the RecruitNet database (Verdú et al. 2023,
\doi{doi:10.1002/ecy.3923}).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_RN()
\dontshow{\}) # examplesIf}
}
