% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtopol.R
\name{funtopol}
\alias{funtopol}
\title{Functional topology of a recruitment network.}
\usage{
funtopol(int_data)
}
\arguments{
\item{int_data}{Data frame with columns named "Canopy" and "Recruit" identifying
observed interactions (i.e. interactions with frequency > 0).}
}
\value{
The function returns two outputs: a data frame with the numeric summary of the
functional structure and a list containing the lists of species of each functional
type (core, satellite, strict transients and disturbance-dependent transients).
}
\description{
Alcantara & Rey (2012) derived a qualitative way to infer potential species
persistence from their position in the RNs based on a combination of non-negative
matrix theory and graph theory. Basically, directed unipartite graphs can be
unambiguously dissected into 'SCCs'. SCCs are the largest possible subgroups of
nodes connected so that all the nodes in a subgroup can be reached from all others
following the directions of the links. In the case of RNs, we can define five
types of SCCs which play different functional roles:
\itemize{
\item The "core" of the network is the SCC formed by the largest number of species.
All species in the core must recruit, at least, in the vicinity of another core
species, and allow the recruitment of at least another core species.
\item “Satellites” are non-core species that can be reached from some core species,
following the direction of the arrows. For example, a satellite species is one
that recruits in the vicinity of some core species but that does not show
recruitment of any species in its vicinity.
\item "Disturbance-dependent transients" are species that can be reached from the open
node but not from core or satellite species (i.e. for example, species that only
recruit away from established plants).
\item "Strict transients" are species that cannot be reached from any other node (i.e.
those that do not recruit in the studied local assemblage).
}

Assuming that the dynamics of the system is linear and time-invariant (LTI dynamics),
like in Markov models (Horn, 1975, Siles et al., 2008), only core and satellite
species will persist in equilibrium in the absence of disturbance. When the dynamics
are non-linear, this result cannot be guaranteed to hold, but simulations have shown
that the probability of persistence is higher and the time to extinction is longer
for core and satellite than for transient species (Alcantara et al., 2017). The sum
of core and satellite species is a qualitative approximation to the number of species
that can potentially persist.
}
\examples{
data(RecruitNet)
int_data <- comm_subset(RecruitNet, site = "Ventisquero")
Ventisquero_funtopol <- funtopol(int_data)
Ventisquero_funtopol$Descriptors
Ventisquero_funtopol$Functional_classification
}
