% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtopol_rec.R
\name{funtopol_rec}
\alias{funtopol_rec}
\title{Species functional role in the general recruitment network}
\usage{
funtopol_rec(int_data, cover_data)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}
}
\value{
A list with two elements, one with descriptors of the networks
and another with the role of each species.

The first element of the list provides the following information:
\itemize{
\item \emph{Num. nodes}: Total number of nodes in the network.
\item \emph{Num. edges}: Total number of links in the network.
\item \emph{Connectance}:  Proportion of links observed relative to all the possible
links (\emph{C}). In the case of recruitment networks, we use the
formula $C = L /(N^2 - N)$ since the node "Open" does not act as a recruit
(i.e. Open is represented by a row of zeroes in the adjacency matrix).
\item \emph{Num. non-trivial SCCs}: Number of SCCs formed by more than one node
in the network.
\item \emph{Num. core species}: Number of species in the core SCC.
\item \emph{Prop. core species}: Proportion of species in the core relative to the
total number of species in the network.
\item \emph{Num. satellite species}: Number of satellite species in the network.
\item \emph{Prop. satellite species}: Proportion of satellite species relative to
the total number of species in the network.
\item \emph{Num. disturbance-dependent transients}: Number of disturbance-dependent
transient species in the network.
\item \emph{Prop. disturbance-dependent transients}: Proportion of
disturbance-dependent transient species relative to the total number of
species in the network.
\item \emph{Num. strict transients}: Number of strict transient species in the network.
\item \emph{Prop. strict transients}: Proportion of strict transient species
relative to the total number of species in the network.
\item \emph{Qualitative Persistence}: The sum of the proportion of core and
satellite species.
The second element of the list, provides the role for each plant species
in the network, whose interpretations differ in each plant-plant interaction
type. Each node (i.e. species) can be:
\item Part of the core \strong{core}. In the case of \emph{general recruitment networks},
the direction of the links represents that the space occupied by canopy
species (predecessor node) will be replaced by the species recruiting
under its canopy (successor node). In this case, all species in the core
must recruit, at least, under the canopy of another core species, and
allow the recruitment of at least another core species.
\item A \strong{satellite} species: these are non-core species that can be reached
from some core species, following the direction of the arrows.
In \emph{general recruitment networks}, these are species that recruit under
the canopy of some core species but that do not show recruitment of any
core species under their canopy.
\item A \strong{disturbance-dependent transient} species: these are species that
can be reached from the "Open" node but not from core or satellite species
(i.e. for example, pioneer species that only recruit away from
established plants).
\item A \strong{strict transient} species: a species that cannot be reached from
any other node. In the case of \emph{general recruitment networks} strict
transients are those species that do not recruit in the studied local
assemblage (e.g., a pioneer species in a mature forest patch or a relict
species at it distribution limit).
}
}
\description{
provides summary information of the frequency of species of
each functional role and the classification of species in each role.
}
\examples{
funtopol_rec(Amoladeras_int,Amoladeras_cover)



}
