% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_db.R
\name{inter_db}
\alias{inter_db}
\title{Function inter_db}
\usage{
inter_db(data, dbcover)
}
\arguments{
\item{data}{a data frame with interactions of recruit & canopy species in each plot of many study sites}

\item{dbcover}{a data frame with the cover of each canopy species per plot in each study site}
}
\value{
a data frame with interactions of recruit & canopy species in each study site and the cover of each species in columns
}
\description{
Function inter_db
}
