% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intertest.R
\name{intertest}
\alias{intertest}
\title{Association tests of interactions}
\usage{
intertest(db_inter, iteration = 100, threshold = 5)
}
\arguments{
\item{db_inter}{data frame with at leas XX columns called XX}

\item{iteration}{number of iterations}

\item{threshold}{minimun number of recruits to perform the binomial test}
}
\value{
data frame with the tests of association
}
\description{
Association tests of interactions
}
