% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_details.R
\name{lc_details}
\alias{lc_details}
\title{TODO: title}
\usage{
lc_details(dataset)
}
\arguments{
\item{dataset}{A data frame for a single site (or a group of sites).
For example, a data frame created with local_comm. This function relies on
the variable names used in the original RecruitNet database, so it cannot be used
with other datasets unless they follow exactly the ordering and the names used in RecruitNet.}
}
\value{
A data frame with summary information of the local community/ies.
}
\description{
This function provides a summary of the characteristics (i.e. metadata) of
a local community.
}
\examples{
#Ventisquero_details <- lc_details(Ventisquero_raw)
#All_sites_details <- lc_details(RecruitNet)
}
