% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_completeness.R
\name{link_completeness}
\alias{link_completeness}
\title{Estimates the completeness and coverage of the sampled interactions}
\usage{
link_completeness(int_data = NULL, type = c("incidence", "abundance"))
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{type}{indicates whether the completeness and coverage will be
calculated based on the number of individual recruits under each canopy
species (i.e., abundance) or on the presence/absence of canopy-recruit pairs
across plots (i.e., incidence). If the survey was based in multiple
independent plots, it is good practice to use the "incidence" approach. The
"abundance" approach can be used more generally, but it assumes that each
recruit represents an independent canopy-recruit interaction event, so it may
easily overestimate coverage when recruits of the same species tend to occur
aggregated under the same canopy plant (as it occurs frequently).
Explanation of its options:
\itemize{
\item \strong{incidence}: Uses the incidence of a given interaction as number of plots
where it was observed.
\item \strong{abundance}: Uses the abundance of a given interaction as the number of
recruits by canopy-recruit.
}}
}
\value{
A data frame indicating the number of links observed (Lobs), the
number of links estimated (Lest), the completeness of the sampling of links
as the proportion of canopy-recruit pairs observed from those expected in
the study site (Completeness of links: Lobs/Lest) and the probability that
if one more recruit (or plot) had been sampled, it would have corresponded
to (or contained) an already detected canopy-recruit species interaction
(Coverage of links).
}
\description{
Estimates the completeness and coverage of the sampled interactions,
providing the probability of detecting a new canopy-recruit pair if we
increased our sampling effort in one unit (i.e., locating one more recruit
or surveying one more plot). This function uses \code{\link[iNEXT:iNEXT]{iNEXT::iNEXT()}}
(Hsieh et al. 2016).
}
\examples{
link_completeness(Amoladeras_int, type="abundance")
link_completeness(Amoladeras_int, type="incidence")

}
