% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cover.R
\name{load_cover}
\alias{load_cover}
\title{Load canopy cover data from RecruitNet database}
\usage{
load_cover(path = getwd())
}
\arguments{
\item{path}{Path to folder containing the 'CanopyCover.csv' file. If the file
is not present, the latest version of the RecruitNet database (Verdú et al. 2023,
\doi{doi:10.1002/ecy.3923}) will be downloaded automatically.}
}
\value{
a data frame containing canopy cover data
}
\description{
Load canopy cover data from RecruitNet database
}
\examples{
cover_data <- load_cover()
}
