% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interactions.R
\name{load_interactions}
\alias{load_interactions}
\title{Load interaction data from RecruitNet database}
\usage{
load_interactions(path = getwd())
}
\arguments{
\item{path}{Path to folder containing the 'RecruitNet.csv' file. If the file
is not present, the latest version of the RecruitNet database (Verdú et al. 2023,
\doi{doi:10.1002/ecy.3923}) will be downloaded automatically.}
}
\value{
a data frame containing interaction data
}
\description{
Load interaction data from RecruitNet database
}
\examples{
int_data <- load_interactions()
}
