% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_RN_cover.R
\name{merge_RN_cover}
\alias{merge_RN_cover}
\title{Function merge_RN_cover}
\usage{
merge_RN_cover(RN_data, cover_data)
}
\arguments{
\item{RN_data}{A data frame created with local_comm, containing all the
available data from a local community.}

\item{cover_data}{data set where cover data are stored. Usually, it will be "CanopyCover"
(or the name assigned to the dataset "CanopyCover.csv" when it was imported).}
}
\value{
A data.frame containing 5 columns with all the information needed
for the basic analysis of recruitment networks and canopy-recruit interactions:
canopy species (canopy), recruit species (recruit), recruitment frequency (fij),
cover of the canopy (cj) and cover of the recruit (ci).
}
\description{
This function makes a data frame that merges the information from the recruitment network
and species cover for a local community.
}
\examples{
#Ventisquero_RNc <- merge_RN_cover(Ventisquero_RN, Ventisquero_cover)
}
