% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacion.R
\name{multiply}
\alias{multiply}
\title{Multiplication}
\usage{
multiply(x = NULL, y = NULL)
}
\arguments{
\item{x}{First Number}

\item{y}{Second Number}
}
\value{
Integer
}
\description{
Multiplication
}
\examples{
multiply(2,3)
}
