% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_degrees.R
\name{node_degrees}
\alias{node_degrees}
\title{Calculate width of canopy service and recruitment niches}
\usage{
node_degrees(int_data, cover_data, int_type = c("rec", "fac", "comp"))
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{int_type}{Indicates the type of plant-plant interaction that will be presented in the
output matrix: general recruitment, recruitment enhancement (i.e.
facilitation) or recruitment depression (i.e. competition). It may take three
possible values:
\itemize{
\item \emph{rec}: Estimates the node degree based on all plant-plant interaction
that contribute to recruitment.
\item \emph{fac}: Estimates the node degree based on only those pairwise
interactions that enhance recruitment. Not every observed interaction in
the field has to be present in the matrix. In this case non-detected
interactions are not considered and "Open" is not included as a canopy
species category.
\item \emph{comp}: Estimates the node degree based on only those pairwise
interactions that depress recruitment. Not every observed interaction in
the field has to be present in the matrix. However, in this case
non-detected interactions are considered (i.e. expanding with 0 all
possible interactions in the study system). "Open" is not included as a
canopy species category.
}}
}
\value{
The output depends on the type of interaction considered.
For the recruitment networks it returns a single data frame with the
following information:
\itemize{
\item \emph{Node}: plant species (either as canopy or recruit)
\item \emph{Ac}: Area of the canopy species (and "Open" interspaces), measured as the
distance (m, in transects) or area (m^2^, in plots) used to estimate canopy
cover in the study site.
\item \emph{canopy_service_width}: the number of species that recruit under a given
canopy species (i.e., the canopy species' out-degree); this can be
interpreted as the width of its canopy service.
\item \emph{canopy_contribution}: number of recruits of any species associated with
the canopy species. When the degrees are weighted by the frequency of
recruitment (Fcr), then it can be interpreted as the contribution of the
canopy species to the multispecific sapling bank.
\item \emph{effective_canopy_service}: The weighted degrees are transformed to the
effective number of partners, accounting for the dominance of certain
interactions in the recruit bank. The effective canopy service is estimated
as e^H, where \emph{H} is Shannon's index of diversity calculated for a canopy
species from Fcr.
\item \emph{recruitment_niche_width}: number of canopy species that allow its
recruitment (the recruit species' in-degree); this can be interpreted as the
width of its recruitment niche.
\item \emph{recruit_bank_abundance}: number of recruits of a species in the study site.
When the degrees are weighted by the frequency of recruitment (Fcr), then it
can be interpreted as species abundance in the recruit bank.
\item \emph{effective_recruitment_niche}: The weighted degrees are transformed to the
effective number of partners, accounting for the dominance of certain
interactions in the recruit bank. The effective recruitment niche width is
estimated as $e^H$, where \emph{H} is Shannon's index of diversity calculated for
a recruit species from Fcr.
}

For facilitation or recruitment depressing interactions, the output consists
in two data frames, one for the  \strong{canopy} species and another for the
\strong{recruit} species. The data frame with information for the canopy species
provides the following information:
\itemize{
\item \emph{Nurse (or Canopy) sp}: Canopy species.
\item \emph{Ac}: Area of the canopy species (and "Open" interspaces), measured as the
distance (m, in transects) or area (m^2^, in plots) used to estimate canopy
cover in the study site.
\item \emph{N enhanced (or depressed) recruit sp}: number of species whose recruitment
is enhanced (or depressed) under its canopy, this can be interpreted as the
width of its enhancing of (depressing) effect.
}

The data frame with information for the recruit species contains the
following information:
\itemize{
\item \emph{Facilitated (or recruit) sp}: Recruit species.
\item \emph{N_ind}: number of recruits of the facilitated (or depressed) species in
the study site.
\item \emph{N enhancing (or depressing) canopy sp}: number of canopy species that
enhance its recruitment (i.e. nurse species) or depress it (i.e. competing
species). These can be interpreted as the width of its niche of nurses or
competitors.
}
}
\description{
Calculate width of canopy service and recruitment niches
}
\examples{
out <- node_degrees(Amoladeras_int, Amoladeras_cover, int_type="rec")
head(out)
}
