% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_com2.R
\name{plot_to_com2}
\alias{plot_to_com2}
\title{TODO: title}
\usage{
plot_to_com2(inter_data = RecruitNet, cover_data = CanopyCover)
}
\arguments{
\item{inter_data}{data frame with at least four columns:
Study_site (unique name of the study site),
Recruit (species of the recruit), Canopy (species of the canopy, or "Open"),
and Frequency (number of recruits of that species observed under that
canopy species in any plot, with "Open" representing recruits observed
without a canopy species).}

\item{cover_data}{data frame with at least five columns:
Study_site (unique name of the study site),
Plot (unique name of each plot within a study site),
Canopy (species of the canopy),
Cover (percentage of cover of that species in that plot), and
Sampled_distance_or_area (total area of that plot or length in the case of transects).}
}
\value{
a list with two elements. The first element is a data frame with
the same structure as the input inter_data, but with two additional columns:
inter_ID (a unique identifier for each pair-wise interaction,
combining Study_site, Recruit, and Canopy species) and
Freq (the number of recruits of that species under that canopy species in
the entire study site).
The second element is a data frame with one row per canopy species in
each Study_site, including two new variables:
Canopy_cover (the percentage of cover of that species in the entire
sampled area of the Study_site) and mycover_sp (a unique identifier
for each canopy species in each Study_site, combining Study_site and Canopy species).
}
\description{
Convert plots into community data per site (either for a single or multiple sites)
}
\examples{
#all_nets <- plot_to_com(RecruitNet, CanopyCover)
}
