% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_niche_test.R
\name{recruitment_niche_test}
\alias{recruitment_niche_test}
\title{Recruitment niche: community-level effects on species recruitment}
\usage{
recruitment_niche_test(int_data, cover_data)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}
}
\value{
The same variables provided by the function \code{\link[=int_significance]{int_significance()}}
but without distinguishing between canopy species.
}
\description{
tests whether the presence of vegetation (i.e any canopy species) compared to the
"Open", enhances, depresses or has a neutral effect on the recruitment of a given
recruit species. \strong{Input}:The canopy-recruit interactions dataset and the canopy
cover dataset (details as explained in \code{\link[=int_significance]{int_significance()}} documentation).
}
\examples{
recruitment_niche_test (Amoladeras_int, Amoladeras_cover)

}
