% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitnetP-package.R
\docType{package}
\name{RecruitNetP-package}
\alias{RecruitNetP}
\alias{RecruitNetP-package}
\title{RecruitNetP: Analysis of Recruitment Networks}
\description{
Analyze recruitment networks and patterns of plant-plant interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/RecruitNetP}
  \item Report bugs at \url{https://github.com/EcologyR/RecruitNetP/issues}
}

}
\author{
\strong{Maintainer}: Alicia Montesinos-Navarro \email{alimontesinos@gmail.com} (\href{https://orcid.org/0000}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item José Luis Garrido \email{jlgarridosan@gmail.com} (\href{https://orcid.org/0000}{ORCID}) [copyright holder]
  \item Miguel Verdú \email{Miguel.Verdu@ext.uv.es} (\href{https://orcid.org/0000}{ORCID}) [copyright holder]
  \item Julio M. Alcántara \email{jmalcan@ujaen.es} (\href{https://orcid.org/0000}{ORCID}) [copyright holder]
  \item Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
}

}
\keyword{internal}
