% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_no_cover.R
\name{remove_no_cover}
\alias{remove_no_cover}
\title{Remove species with no cover data}
\usage{
remove_no_cover(
  int_data,
  cover_data,
  rm_sp_no_cover = c("allsp", "onlycanopy")
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{rm_sp_no_cover}{Options:
\itemize{
\item \emph{allsp}: removes any species without cover data.
\item \emph{onlycanopy}: removes only canopy species without cover.
}}
}
\value{
a data frame
}
\description{
Remove the species from the interaction database that do not have cover data. For rec
interaction type the function removes any species without cover data from the
interactions database, and for comp or fac, it removes only canopy species without
cover data keeping recruit species without cover data.
}
\examples{
remove_no_cover (Amoladeras_int, Amoladeras_cover, rm_sp_no_cover = "allsp")
remove_no_cover (Amoladeras_int, Amoladeras_cover, rm_sp_no_cover = "onlycanopy")

}
