% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RN}
\alias{RN}
\title{RecruitNet}
\format{
An object of class \code{data.frame} with 135210 rows and 26 columns.
}
\source{
https://doi.org/10.1002/ecy.3923
}
\usage{
RN
}
\description{
Plant-plant interactions global database
}
\keyword{datasets}
