% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_heatmap.R
\name{RN_heatmap}
\alias{RN_heatmap}
\title{Visualize a heatmap of the interactions matrix.}
\usage{
RN_heatmap(
  int_data,
  cover_data,
  int_type = c("rec", "fac", "comp"),
  weight = c("Pcr", "Fcr", "Dcr", "Dro", "Ns", "NintC", "NintA", "RII")
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{int_type}{#' Indicates the type of plant-plant interaction that will be presented in the
output matrix: general recruitment, recruitment enhancement (i.e.
facilitation) or recruitment depression (i.e. competition). It may take three
possible values:
\itemize{
\item \emph{rec}: Estimates the node degree based on all plant-plant interaction
that contribute to recruitment.
\item \emph{fac}: Estimates the node degree based on only those pairwise
interactions that enhance recruitment. Not every observed interaction in
the field has to be present in the matrix. In this case non-detected
interactions are not considered and "Open" is not included as a canopy
species category.
\item \emph{comp}: Estimates the node degree based on only those pairwise
interactions that depress recruitment. Not every observed interaction in
the field has to be present in the matrix. However, in this case
non-detected interactions are considered (i.e. expanding with 0 all
possible interactions in the study system). "Open" is not included as a
canopy species category.
}}

\item{weight}{Specifies the metric used to represent interaction strength (i.e., the
weight) assigned to each pair of species in the matrix. See function
\strong{associndex} for further details. The possible options are:
\itemize{
\item \emph{Fcr}: \strong{frequency of recruitment} in number of recruits by canopy-recruit
pair.
\item \emph{Pcr}: \strong{presence/absence} (1/0) of the interaction (unweighted network).
\item \emph{Dcr}: \strong{density of recruitment} as number of recruits per unit area of
canopy species.
\item \emph{Dro}: \strong{density of recruitment in open interspaces} as number of
recruits per unit area of open interspaces.
\item \emph{Ns}: The index \strong{Normalized Neighbour Suitability index} (proposed by
Mingo, 2014), suitable for comparisons of interaction strength between pairs
of species within a local community, which should be preferred in general
recruitment networks (Alcantara et al. 2025).
\item \emph{NIntA}: The index \strong{additive symmetry intensity index} proposed by
Diaz-Sierra et al. (2017).
\item \emph{NIntC}: The index \strong{commutative symmetry intensity index} proposed by
Diaz-Sierra et al. (2017).
\item \emph{RII}: The index \strong{Relative Interaction Index} (Armas et al., 2004).
}}
}
\value{
Heatmap plot of the weighted network.
}
\description{
Visualize a heatmap of the interactions matrix.
}
\examples{
RN_heatmap(Amoladeras_int, Amoladeras_cover, int_type="fac", weight ="Ns")

}
