% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtest.R
\name{sigtest}
\alias{sigtest}
\title{TODO: name}
\usage{
sigtest(data_freq, iteration = 100, threshold = 5)
}
\arguments{
\item{data_freq}{data frame obtained as the output of the functions pre_asocindex,
recruit_level,canopy_level, or com_level}

\item{iteration}{number of iterations}

\item{threshold}{minimun number of recruits to perform the binomial test}
}
\value{
data frame with the same structure as the input with three additional columns:
int_p (p-value of the binomial test of association),
int_sign (the sign of the association, being Posive (or Negative) if the
association is stronger (or weaker) than expected by the percentage cover of Canopy and Open,
and Neutral if there is not enougth power to conduct the test,
stdres (standarized resdual quantifying the difference between the observed and expected values),
and testability (indicating whether the sample size allow to conduct or not the test
being Non-testable those rows in which int_sign = Neutral)
}
\description{
Calculate a significance tests of association for each row.
Rows can be a pair-wise interaction (Recruit species-Canopy species interaction),
the number of recruits observed under any canopy species (Recruit species)
or the number of Recruits of any given speies under a given canopy species (Canopy species)
}
\examples{
#sigtest(com_level(RecruitNet, CanopyCover))
#sigtest(pre_asocindex(RecruitNet, CanopyCover)) # "very slow"
}
