% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_local_comm.R
\name{split_local_comm}
\alias{split_local_comm}
\title{Function split_local_comm}
\usage{
split_local_comm(raw_RN)
}
\arguments{
\item{raw_RN}{A data frame containing all the data from a community sampled
in multiple plots. It must contain, at least, the columns "Plot", "Canopy",
"Recruit" and "Frequency".}
}
\value{
A list of data frames, each containing the frequency of
canopy-recruit interactions in a plot.
}
\description{
When a local community has been sampled in multiple plots, we may need to
work with data from each plot separately. This function produces a list with
the basic data to build a RN for each plot.
}
\examples{
#Ventisquero_plots <- split_local_comm(Ventisquero_raw)
}
