% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topol_fac.R
\name{topol_fac}
\alias{topol_fac}
\title{summary and node-based information of species in loops and simple paths}
\usage{
topol_fac(int_data, cover_data, direction = c("in", "out"))
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{direction}{the direction of the link}
}
\value{
A list with two elements, one with the information of reciprocal
facilitation loops and another with information of the simple linear
indirect facilitation paths beginning of ending in each species.

The first element of the list \strong{loops} provides the following information
within two levels:
\itemize{
\item \strong{summary}: A data frame with as many rows as SCC are present in the
facilitation network (with more than one species, as autofacilitation is
not considered) and in columns, an scc identifier \strong{scc_id} and the number
of species involved in that SCC \strong{n_nodos}
\item \strong{nodes}: a list with as many elements as SCCs found, each of them
including a vector with the name of the species involved in that SCC
The second element of the list \strong{simple}, provides the following
information within two levels:
\item \strong{summary}: A data frame with as many rows as distinct simple
linear indirect facilitation paths (i.e. their overlap in species
composition is less than 75\%) that begin ("from") or end ("to") in any
species, including the linear versions of the SCCs. In columns, the species
in which that distinct linear path begin or end  \strong{nodo}, and numeric
index that identifies the paths within species \strong{path_index} and the
number of species involved in that path \strong{n_nodes_in_path} which includes
the focal species, independently of whether there is autofacilitation.
\item \strong{nodes}: a list with as many elements as species in the facilitation
network (either as nurse or facilitated), each of them including as many
elements as distinct paths begin or ends in it, and in each of them, a
vector with the name of the species involved in that simple indirect
facilitation linear path.
All arguments (options):
\item \strong{direction} = c("in","out")
Argument 1.
\item \strong{direction}: Indicates the direction in which the indirect facilitation
is assessed.
Explanation of its options:
\item \strong{in}: Estimates the paths based on the incoming links to each node,
representing the set of nurse species from which a recruit species
benefits, either directly (by enhancing its recruitment) or indirectly
(by facilitating other nurse species that enhance its recruitment).
Link direction is considered for both reciprocal and simple paths.
\item \strong{out}: Estimates the paths based on the outgoing links from each nodes,
representing the set of species that a nurse species benefits, either
directly or indirectly (i.e. through other intermediate species).
The direction of the links is applied to both reciprocal and simple paths
}
}
\description{
provides summary and node-based information of number of species involved
in loops of reciprocal facilitation or simple indirect facilitation linear
paths from (or to) any given species.
}
\examples{

topol_fac(test_data$com,test_data$cov, direction="out")

topol_fac(test_data$com,test_data$cov, direction="in")


}
