% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_funtopol_rec.R
\name{visu_funtopol_rec}
\alias{visu_funtopol_rec}
\title{visualize the functional topology of general recruitment networks}
\usage{
visu_funtopol_rec(int_data, cover_data)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}
}
\value{
a plot with teh different species role
}
\description{
visualize the functional topology of general recruitment networks
}
\examples{

visu_funtopol_rec(Amoladeras_int,Amoladeras_cover)

}
