% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_topol_depre.R
\name{visu_topol_depre}
\alias{visu_topol_depre}
\title{Allow the visualization of intransitivity loops of recruitment depression}
\usage{
visu_topol_depre(
  int_data,
  cover_data,
  layout_fun = igraph::layout_with_fr,
  vertex_size = 20,
  edge_arrow_size = 0.4,
  direction = c("out", "in")
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{layout_fun}{type of igraph layout (see \code{\link[igraph:layout_]{igraph::layout_()}})}

\item{vertex_size}{numeric}

\item{edge_arrow_size}{numeric}

\item{direction}{Direction of the displayed links, either "out" or "in".}
}
\value{
a plot
}
\description{
with nodes in different SCC identified by different colors (in grey if they
do not belong to any SCC).
It may be frequent that in recruitment enhancement (i.e. facilitation) or
depression (i.e.competition) networks, there is not any SCC as they tend to
have smaller dimensions that general recruitment networks.
}
\examples{
visu_topol_depre(test_data$com,test_data$cov)

}
