% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associndex.R
\name{associndex}
\alias{associndex}
\title{Calculate different association indices considering different interaction types.}
\usage{
associndex(
  int_data,
  cover_data,
  expand = c("yes", "no"),
  rm_sp_no_cover = c("allsp", "onlycanopy"),
  threshold_density = NULL
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{expand}{Expands the data frame to include the non-detected interactions
between every pair of species that occur in the study site, assigning values
of 0 to the variables Fcr and Dcr. Explanation of its options:
\itemize{
\item \emph{yes}: Generates all possible interactions between canopy and recruit species
in the study site (i.e. detected and non-detected), adding values of interaction
frequency of 0 to the non-detected interactions.
\item \emph{no}: Considers only the pairwise interactions actually detected in the study site.
}}

\item{rm_sp_no_cover}{\strong{rm_sp_no_cover}: Defines the filtering behavior, specifying
which species should be removed based on the availability of cover data and the type
of interaction of interest. Explanation of its options:
\itemize{
\item \emph{allsp}: removes any species that lacks cover data (both canopy and recruit species).
\item \emph{onlycanopy}: removes only canopy species lacking cover data, retaining recruit
species even if they lack cover data.
}}

\item{threshold_density}{\strong{threshold_density}: indicates a threshold to retain only
those interactions in which the observed density is below the threshold
(i.e. Dcr < thr & Dro < thr). This is required to avoid the use of interactions
based on very few recruits which may have occurred by chance under a very scarce
canopy species. These cases will result in a large overestimation of Dcr, likely
being spurious outlayers. If no threshold value is provided or it is set to NULL,
the function estimates a threshold value based on recruitment density outliers
according to a Weibull distribution: a value is an outlier if it is above the limit
where less than 1 observation is expected. Explanation of its options:
\itemize{
\item A positive value. High values provide more chance to include potentially spurious
interactions, unless such a high density is reliable based on the knowledge of the
study system. A way to decide the threshold value would consist in identifying possible
outlayers in the distribution of recruitment densities (\emph{Dcr} and \emph{Dro}).
\item NULL: use package \strong{extremevalues} to find the density threshold value (\emph{K})
above which, according to the Weibull distribution, less than 1 observation is
expected given the number of observations in the sample.
}}
}
\value{
a data frame with the following information for each canopy-recruit
interaction with the following information:
\itemize{
\item \emph{Canopy}: \strong{canopy species}.
\item \emph{Recruit}: \strong{recruit species}.
\item \emph{Fcr}: \strong{frequency of recruitment}, as the number of recruits found
under that canopy species.
\item \emph{Ac}: \strong{Area occupied by the canopy species} in the study site,
in m^2^ (or m when relative cover is measured with transects).
\item \emph{Fro}: \strong{frequency of recruitment}, as the number of recruits in
Open interspaces. - \emph{Ao}: \strong{Area  occupied by open interspaces} in the study
site in m^2^ (or m when relative cover is measured with transects).
\item \emph{Dcr}: \strong{Density of recruits under a canopy species}: $Fcr/Ac$.
\item \emph{Dro}: \strong{Density of recruits in open interspaces}: $Fro/Ao$.
\item \emph{max_Recr}: \strong{maximum recruitment density} of a species in the study site
(i.e., under any canopy species or "Open").
\item \emph{Ns}: \strong{Normalized Neighbour Suitability} index: $(Dcr - Dro) / maxRecr$.
\item \emph{NintC}: \strong{Commutative symmetry intensity} index: $2*(Dcr - Dro)/((Dcr + Dro)+abs(Dcr-Dro))$
\item \emph{NintA}: \strong{Additive symmetry intensity} index: $2*(Dcr - Dro)/((Dro) + abs(Dcr-Dro))$
\item \emph{RII}: \strong{Relative Interaction Index}: $(Dcr - Dro)/(Dcr + Dro)$
}
}
\description{
Calculates different metrics to characterize the interaction strength
for each pairwise interaction, based on the number of recruits found under
the canopy species and in the "Open" ground. Several indices have been proposed
to assess how much recruitment is affected by the canopy species relative
to the species performance in the absence of canopy plants (i.e. in open),
each of them with its own particularities. In the context of recruitment
interactions, the most frequently used index is the \emph{Relative Interaction Index}
(\strong{RII}; Armas et al., 2004), and some studies have recently used the
\emph{additive and commutative symmetry intensity indices} (\strong{NIntA} and \strong{NIntC})
y Diaz-Sierra et al. (2017). However, these indices are not well suited for
comparisons of interaction strength between pairs of species within a local
community, so the \emph{Normalized Neighbour Suitability} index (\strong{Ns}; Mingo, 2014)
should be preferred (Alcantara et al. 2025). The function \emph{\code{associndex}}
calculates all the four indices besides providing the density of recruits
under the canopy and in the open, and the data used to calculate it
(relative cover and number of recruits in each microhabitat).
**IMPORTANT NOTE: Data on recruitment in Open is required to calculate
these indices.
\strong{Input}: The canopy-recruit interactions data set and the canopy cover
data set
}
\examples{
associndex (Amoladeras_int, Amoladeras_cover, expand = "yes",
rm_sp_no_cover = "allsp", threshold_density=NULL)
associndex (Amoladeras_int, Amoladeras_cover, expand = "yes",
rm_sp_no_cover = "onlycanopy", threshold_density=NULL)
associndex (Amoladeras_int, Amoladeras_cover, expand = "no",
rm_sp_no_cover = "allsp", threshold_density=NULL)
associndex (Amoladeras_int, Amoladeras_cover, expand = "yes",
rm_sp_no_cover = "onlycanopy", threshold_density=NULL)
}
