% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_subset.R
\name{comm_subset}
\alias{comm_subset}
\title{Subset sites}
\usage{
comm_subset(dataset = NULL, site = NULL)
}
\arguments{
\item{dataset}{data set where recruitment networks are stored.
When using the RecruitNet database, this name will be "RecruitNet",
but can be any name assigned to the dataset when it is imported.}

\item{site}{name of a study site within the dataset or a character vector
with the names of several study sites.}
}
\value{
A data frame containing all the data from the desired community/ies
}
\description{
This function extracts the information of a given local community and
stores it in a separate data frame. It can also extract more than one.
}
\examples{
#Ventisquero_raw <- comm_subset(RecruitNet, "Ventisquero")

#TwoSites <- c("Ventisquero","Agadir")
#TwoSites_raw <- comm_subset(RecruitNet, TwoSites)
}
