% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_subset.R
\name{comm_subset}
\alias{comm_subset}
\title{Subset data from one study site}
\usage{
comm_subset(data = NULL, site = NULL)
}
\arguments{
\item{data}{a dataframe. Must contain a column named Study_site}

\item{site}{A character string indicating the name of the study site}
}
\value{
a dataframe
}
\description{
this function is to be used with datasets containing multiple study sites
}
\examples{

data<-load_interactions()
comm_subset(data, site="Amoladeras")

}
