% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_completeness.R
\name{link_completeness}
\alias{link_completeness}
\title{Link_completeness}
\usage{
link_completeness(dataset, site, type)
}
\arguments{
\item{dataset}{The name of the data set where recruitment networks are stored.
Usually, it will be "RecruitNet" (or the name assigned to the data set when
it was imported). At least, the data set must contain columns named
"Study_site", "Plot", "Canopy", "Recruit" and "Frequency".
In data sets containing a single site or a single plot,
the corresponding columns must be included anyway.}

\item{site}{The name of a study site.}

\item{type}{Can take two values: "incidence" or "abundance".
Estimates based on incidence data are recommended, but can only be obtained
if your data set is structured in multiple plots. Alternatively, for data
collected in a single plot, estimates can be based only on abundance data.}
}
\value{
A data frame
}
\description{
Observed and estimated number of links, and link completeness.
Methods based in Chao et al. (2014) as implemented in R package iNEXT.
}
\examples{
#link_completeness(RecruitNet, "Ventisquero", "incidence")
#link_completeness(RecruitNet, "Laxe", "abundance")
#link_completeness(RecruitNet, "Laxe", "incidence") # Issues a warning
#link_completeness(RecruitNet, "LosReyes", "abundance")
#link_completeness(RecruitNet, "LosReyes", "incidence") #Issues an error
}
