% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_com.R
\name{plot_to_com}
\alias{plot_to_com}
\title{Function plot_to_com.It summarize the information of many plots per site into a single data frame at the community level}
\usage{
plot_to_com(data, dbcover)
}
\arguments{
\item{data}{a data frame with frequency of recruits under canopy species species in each plot of many study sites}

\item{Canopy_all}{a data frame with the cover of each canopy species per plot in each study site}
}
\value{
a list with three elements: a data frame with the number of recruits observed under each canopy species(including the open), a data frame with the relative cover of each canopy species and Open in each study site and a list of quantitative community matrices (per study site) with recruit and canopy species in rows and columns respectively
}
\description{
Function plot_to_com.It summarize the information of many plots per site into a single data frame at the community level
}
