% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitnetP-package.R
\docType{package}
\name{RecruitnetP-package}
\alias{RecruitnetP}
\alias{RecruitnetP-package}
\title{RecruitnetP: Metrics of recruitment networks}
\description{
Package to analyze patterns of plant-plant interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/templateRpackage}
  \item Report bugs at \url{https://github.com/EcologyR/templateRpackage/issues}
}

}
\author{
\strong{Maintainer}: José Luis Garrido Sánchez \email{jlgarridosan@gmail.com} (\href{https://orcid.org/0000}{ORCID})

Authors:
\itemize{
  \item Alicia Montesinos-Navarro \email{alimontesinos@gmail.com} (\href{https://orcid.org/0000}{ORCID})
  \item Miguel Verdú del Campo \email{Miguel.Verdu@ext.uv.es} (\href{https://orcid.org/0000}{ORCID})
  \item Julio M Alcántara Gámez \email{jmalcan@ujaen.es} (\href{https://orcid.org/0000}{ORCID})
}

}
\keyword{internal}
