% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RII_index.R
\name{RIIindex}
\alias{RIIindex}
\title{Function RII_index.It generates the relative association index described by Armas et al. XXX
#'}
\usage{
RIIindex(db_inter)
}
\arguments{
\item{db_inter}{a data frame a data frame which must include at least seven columns named:"Study_site","Recruit","Canopy","Canopy_Freq","Canopy_cover","Open_Freq","Open_cover" with frequency of recruits of each species under each canopy species and in Open,(Canopy_Freq, Open_Freq), and the percentage of cover occupied by each of them (named: Canopy_cover, Open_cover). It can be generated by the function "asocindex".}
}
\value{
the imputed data frame with a new column with the relative association index (RII)
}
\description{
Function RII_index.It generates the relative association index described by Armas et al. XXX
#'
}
