% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RN_to_matrix.R
\name{RN_to_matrix}
\alias{RN_to_matrix}
\title{Convert field data to matrix format}
\usage{
RN_to_matrix(
  int_data,
  cover_data,
  int_type = c("rec", "fac", "comp"),
  weight = c("Fcr", "Dcr", "Dro", "Ns", "NintC", "NintA", "RII")
)
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{int_type}{Indicates the type of plant-plant interaction that will be
analyzed: general recruitment, recruitment enhancement (i.e. facilitation)
or recruitment depression (i.e. competition).
Explanation of its options:
\itemize{
\item \emph{rec}: All the pairwise interactions observed will be in the output.
Focuses on canopy-recruit interactions considering that every recruit
growing under the canopy of another plant may occupy that space in the
future, thus having a potentially positive effect on the recruit species
population. Therefore, even a single observation is considered an
interaction. This type of networks considers every species present in the
study system, whether as a canopy or as a recruit, as a node in the network.
It also includes "Open" as a particular node since some species may recruit
away from established plants. Non-detected interactions are also considered
since zero frequency can provide evidence of a very negative interaction
if the expected frequency under the canopy species is large.
\item \emph{fac}: Only those pairwise interactions that resulted in recruitment
enhancement will be in the output. Focuses on interactions with a
significantly higher recruitment density under canopy  than in "Open"
(i.e. facilitation).  Non-detected interactions are not considered and
"Open" is not included as a node, although its relative cover is considered
as part of the sampling area.
\item \emph{comp}: Only those pairwise interactions that resulted in a recruitment
depression will be in the output. Focuses on interactions with a
significantly lower recruitment density under canopy than "Open"
(i.e. competition). Non-detected interactions are considered (i.e. expanding
with 0 all possible interactions in the study system), as the absence of
recruitment of a species under a given canopy can reflect a particularly
strong depression of recruitment under that canopy species. "Open" is not
included as a node, although its relative cover is considered as part of the
sampling area.
}}

\item{weight}{Specifies the metric used to represent interaction strength
(i.e., the weight) assigned to each pair of species in the matrix. Explanation of its
options (more mathematical information in the description of the function
\code{\link[=associndex]{associndex()}} :
\itemize{
\item \emph{Fcr}: \emph{\strong{frequency of recruitment}}, in number of recruits by
canopy-recruit pair.
\item \emph{Dcr}: \emph{\strong{density of recruitment}}, as number of recruits per unit
area of canopy species cover.
\item \emph{Dro}: \emph{\strong{density of recruitment in open interspaces}}, as number of
recruits per unit area of open interspaces.
\item \emph{Ns}: The \emph{\strong{Normalized Neighbour Suitability index}}; suitable for
comparisons of interaction strength between pairs of species within a
local community.
\item \emph{NIntA}: The \emph{\strong{Additive symmetry intensity index}}.
\item \emph{NIntC}: The \emph{\strong{Commutative symmetry intensity index}}.
\item \emph{RII}: The \emph{\strong{Relative Interaction Index}}.
}}
}
\value{
A matrix with recruit species in rows and canopy species in columns,
with cells describing the measure selected to describe the interaction
strength (i.e. weight) between each pair of species.
}
\description{
converts the community data collected in the field, where each row represents
a single canopy-recruit interaction, into a matrix format, with recruit
species as rows and canopy species as columns. For general recruitment
networks, the matrix should be square including all species observed at the study site. This means that the same set of species appears in both rows and columns, with "Open" areas treated as an additional category in both the Canopy and Recruit variables. In contrast, for recruitment enhancement (i.e. facilitation) or depression (i.e. competition), the matrix can be non-square. In this case, rows represent the species whose recruitment is enhanced or suppressed, while columns represent the canopy species that influence this recruitment. These two groups may not include the same species.
}
\examples{

RN_to_matrix(Amoladeras_int, Amoladeras_cover, int_type="rec", weight="Dcr")


}
