% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topol_depre.R
\name{topol_depre}
\alias{topol_depre}
\title{Summary and node-based information of species in loops and simple paths}
\usage{
topol_depre(int_data, cover_data, direction = c("in", "out"))
}
\arguments{
\item{int_data}{data frame containing interaction data.}

\item{cover_data}{data frame with the abundance of each canopy species in each plot.}

\item{direction}{direction of teh links, in or out}
}
\value{
A list with two elements, one with the information of intransitive
loops of recruitment depression and another with information of the simple
indirect recruitment depression linear paths beginning of ending in
each species.

The first element of the list \strong{loops} provides the following information
within two levels:
\itemize{
\item \strong{summary}: A data frame with as many rows as SCC are present in the
recruitment depression network (with more than one species, as
autodepression is not considered) and in columns, an scc identifier
\strong{scc_id} and the number of species involved in that SCC \strong{n_nodos}
\item \strong{nodes}: a list with as many elements as SCCs found, each of them
including a vector with the name of the species involved in that SCC
The second element of the list \strong{simple}, provides the following
information within two levels:
\item \strong{summary}: A data frame with as many rows as distinct simple linear
indirect recruitment depression paths (i.e. their overlap in species
composition is less than 75\%) that begin ("from") or end ("to") in any
species, including the linear versions of the SCCs. In columns, the species
in which that distinct linear path begin or end  \strong{nodo}, and numeric
index that identifies the paths within species \strong{path_index} and the
number of species involved in that path \strong{n_nodes_in_path} which includes
the focal species, independently of whether there is autodepression.
\item \strong{nodes}: a list with as many elements as species in the recruitment
depression network (either as depresor or depressed), each of them
including as many elements as distinct paths begin or ends in it, and in
each of them, a vector with the name of the species involved in that simple
indirect recruitment depression linear paths.
All arguments (options)**:
\item \strong{direction} = c("in","out")
Argument 1.
\item \strong{direction}: Indicates the direction in which the indirect recruitment
depression is assessed.
Explanation of its options:
\item \strong{in}: Estimates the paths based on the incoming links to each node,
representing the set of canopy species that affect the recruitment of a
given species through depressive effects, either directly (by depressing
its recruitment) or indirectly (by depressing the recruitment of other
canopy species connected to it through recruitment depression interactions).
Link direction is considered for both reciprocal and simple paths.
\item \strong{out}: Estimates the paths based on the outgoing links from each nodes,
representing the set of species that canopy species affect through
depressive effects, either directly or indirectly
(i.e. through other intermediate species). The direction of the links
is applied to both reciprocal and simple paths
}
}
\description{
provides summary and node-based information of number of species involved
in intransitive loops of recruitment depression or simple indirect
recruitment depression linear paths from (or to) any given species.
}
\examples{
topol_depre(test_data$com,test_data$cov, direction="out")

topol_depre(test_data$com,test_data$cov, direction="in")


}
